requirements.txt
gpiozero
opencv-python
python-vlc
vosk
pyaudio
requests
numpy
